Make new Application class
